refactor(db) 4/13: extract db/quest.rs + db/testutil.rs (closes #250)
Move 15 quest/mission/chain/cluster populators (+ migrate_quest_typed_columns),
6 quest classifiers, the quest CREATE TABLEs (into quest::create_tables, wired
into init_schema), and 7 quest tests into db/quest.rs. New db/testutil.rs holds
the shared temp_db_path/seed_singleton harness (pub(crate), #[cfg(test)]).
mod.rs imports re-exported pub(crate) so domain files `use super::*`.
240 tests + clippy -D warnings + fmt green; table-count test confirms no DDL lost.
